% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_latest_ons_data}
\alias{download_latest_ons_data}
\title{Downloads first (assuming most recent) dataset link embedded on ONS webpage}
\usage{
download_latest_ons_data(ons_url, destfile)
}
\arguments{
\item{ons_url}{The URL for the ONS webpage which contains embedded dataset link/s.}

\item{destfile}{The destination location where you want to save the file.}
}
\value{
The destination file path
}
\description{
Downloads first (assuming most recent) dataset link embedded on ONS webpage
}
\examples{
ons_url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19infectionsurveydata"

destfile <- "data/cisdata.xlsx"

download_latest_ons_data(ons_url, destfile)

}
