% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlatestdata.R
\name{download_nhs_data}
\alias{download_nhs_data}
\title{Downloads specified dataset link embedded on NHS webpage
and saves in the user specified destination location (and will create new
folder directories if required)}
\usage{
download_nhs_data(nhs_url, item_number, destfilepath)
}
\arguments{
\item{nhs_url}{The URL for the NHS webpage which contains embedded dataset link/s.}

\item{item_number}{The element number (based on element order in data_urls) of the URL of interest}

\item{destfilepath}{The destination location where you want to save the file.Note that the file format type must match the format of the file you are trying to download.}
}
\value{
The destination file path
}
\description{
Downloads specified dataset link embedded on NHS webpage
and saves in the user specified destination location (and will create new
folder directories if required)
}
\examples{
nhs_url <- paste("https://digital.nhs.uk/data-and-information/publications/",
"statistical/appointments-in-general-practice/june-2023", sep="")

item_number <- 4

destfilepath <- "data.csv"

download_nhs_data(nhs_url, 4, destfilepath)

}
