% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlatestdata.R
\name{get_all_nhs_data_urls}
\alias{get_all_nhs_data_urls}
\title{Finds all embedded data links on an NHS webpage}
\usage{
get_all_nhs_data_urls(nhs_url)
}
\arguments{
\item{nhs_url}{The URL for the NHS webpage which contains embedded dataset link/s.}
}
\value{
data_urls - all data URLs on the webpage
}
\description{
Finds all embedded data links on an NHS webpage
}
\examples{
nhs_url <- paste("https://digital.nhs.uk/data-and-information/publications/",
"statistical/appointments-in-general-practice/june-2023", sep="")
get_all_nhs_data_urls(nhs_url)

}
