% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{get_dataset_title}
\alias{get_dataset_title}
\title{Retrieves dataset title from ONS webpage}
\usage{
get_dataset_title(ons_url)
}
\arguments{
\item{ons_url}{The ONS URL which contains the embedded dataset link.}
}
\value{
dataset_title The title of the dataset.
}
\description{
Retrieves dataset title from ONS webpage
}
\examples{
ons_url <- paste("https://www.ons.gov.uk/peoplepopulationandcommunity/",
"healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19",
"infectionsurveydata", sep="")

get_dataset_title(ons_url)

}
