% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_html_text}
\alias{get_html_text}
\title{Retrieves html text}
\usage{
get_html_text(ons_url, element_info)
}
\arguments{
\item{ons_url}{The ONS URL from which you want to read text.}

\item{element_info}{The html element you want to read.}
}
\value{
html_text_script The text from the element you have specified.
}
\description{
Retrieves html text
}
\examples{
ons_url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19infectionsurveydata"

element_info <- "h1.page-intro__title"

get_html_text(ons_url, element_info)
}
