% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlatestdata.R
\name{get_latest_ons_data_url}
\alias{get_latest_ons_data_url}
\title{Finds the first (assuming most recent) data (xlsx) URL on an ONS webpage}
\usage{
get_latest_ons_data_url(ons_url)
}
\arguments{
\item{ons_url}{The URL for the ONS webpage which contains embedded dataset link/s.}
}
\value{
data_url - the first, assuming most recent, data URL on the webpage
}
\description{
Finds the first (assuming most recent) data (xlsx) URL on an ONS webpage
}
\examples{
ons_url <- paste("https://www.ons.gov.uk/peoplepopulationandcommunity/",
"healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19",
"infectionsurveydata", sep="")

get_latest_ons_data_url(ons_url)

}
