% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Higher level function that retrieves relevant metadata for ONS datasets}
\usage{
get_metadata(ons_url)
}
\arguments{
\item{ons_url}{The ONS URL which contains the embedded dataset link.}
}
\value{
metadata A list containing the dataset title, latest edition, release date, when it will next be updated and additional information about the dataset.
}
\description{
Higher level function that retrieves relevant metadata for ONS datasets
}
\examples{
ons_url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19infectionsurveydata"

get_metadata(ons_url)
}
