% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_next_updated}
\alias{get_next_updated}
\title{Retrieves the date the next version of the dataset will be made available}
\usage{
get_next_updated(ons_url)
}
\arguments{
\item{ons_url}{The ONS URL which contains the embedded dataset link.}
}
\value{
next_update The date the next version of the dataset will be made available
}
\description{
Retrieves the date the next version of the dataset will be made available
}
\examples{
ons_url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19infectionsurveydata"

get_next_updated(ons_url)
}
