% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlatestdata.R
\name{select_nhs_data_url}
\alias{select_nhs_data_url}
\title{Select the data link of interest from all embedded data links on an NHS webpage}
\usage{
select_nhs_data_url(nhs_url, item_number)
}
\arguments{
\item{nhs_url}{The URL for the NHS webpage which contains embedded dataset link/s.}

\item{item_number}{The element number (based on element order in data_urls) of the URL of interest}
}
\value{
data_url - the data URL of interest on the webpage
}
\description{
Select the data link of interest from all embedded data links on an NHS webpage
}
\examples{
nhs_url <- paste("https://digital.nhs.uk/data-and-information/publications/",
"statistical/appointments-in-general-practice/june-2023", sep="")

item_number <- 4
select_nhs_data_url(nhs_url, item_number)

}
