% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{get_available_editions}
\alias{get_available_editions}
\title{Retrieves a list of all available editions of the dataset}
\usage{
get_available_editions(ons_url)
}
\arguments{
\item{ons_url}{The ONS URL which contains the embedded dataset link.}
}
\value{
available_editions A list of all available editions of the dataset.
}
\description{
Retrieves a list of all available editions of the dataset
}
\examples{
ons_url <- paste("https://www.ons.gov.uk/peoplepopulationandcommunity/",
"healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19",
"infectionsurveydata", sep="")

get_available_editions(ons_url)
}
