% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{get_html_text}
\alias{get_html_text}
\title{Retrieves html text}
\usage{
get_html_text(nhs_url, element_info)
}
\arguments{
\item{nhs_url}{The NHS URL from which you want to read text.}

\item{element_info}{The html element you want to read.}
}
\value{
html_text_script The text from the element you have specified.
}
\description{
Retrieves html text
}
\examples{
nhs_url <- paste("https://digital.nhs.uk/data-and-information/publications/",
"statistical/appointments-in-general-practice/june-2023", sep="")

element_info <- "dt.nhsd-o-hero__meta-data-item-title"

get_html_text(nhs_url, element_info)
}
