% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{get_latest_edition}
\alias{get_latest_edition}
\title{Retrieves the latest(first listed) edition name of the dataset}
\usage{
get_latest_edition(ons_url)
}
\arguments{
\item{ons_url}{The ONS URL which contains the embedded dataset link.}
}
\value{
latest_edition The latest edition name of the dataset
}
\description{
Retrieves the latest(first listed) edition name of the dataset
}
\examples{
ons_url <- paste("https://www.ons.gov.uk/peoplepopulationandcommunity/",
"healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19",
"infectionsurveydata", sep="")

get_latest_edition(ons_url)
}
