% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieves metadata from NHS webpage}
\usage{
get_metadata(nhs_url)
}
\arguments{
\item{nhs_url}{The NHS URL which contains the embedded dataset link.}
}
\value{
metadata A table containing the publication title, publication date, geographic coverage and granularity and date range.
}
\description{
Retrieves metadata from NHS webpage
}
\examples{
nhs_url <- paste("https://digital.nhs.uk/data-and-information/publications/",
"statistical/appointments-in-general-practice/june-2023", sep="")

get_metadata(nhs_url)

}
