% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_release_date}
\alias{get_release_date}
\title{Retrieves the release date of the latest dataset available}
\usage{
get_release_date(ons_url)
}
\arguments{
\item{ons_url}{The ONS URL which contains the embedded dataset link.}
}
\value{
release_date The release date of the latest dataset available.
}
\description{
Retrieves the release date of the latest dataset available
}
\examples{
ons_url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19infectionsurveydata"

get_release_date(ons_url)
}
